Add `TinkerComputeQueryRunner`
